fix: move dependency (params) classes from utils to params.py chore: remove empty utils folder
