+++ tests/api_swedeb/api/services/test_ngrams_service.py
@@ -5,8 +5,8 @@ from unittest.mock import MagicMock, Mock, patch
 import pandas as pd
 
 from api_swedeb import schemas
+from api_swedeb.api.params import CommonQueryParams
 from api_swedeb.api.services.ngrams_service import NGramsService
-from api_swedeb.api.utils.common_params import CommonQueryParams
 from api_swedeb.schemas import NGramResult
 
 
@@ -27,8 +27,6 @@ class TestNGramsServiceMethods:
     @patch('api_swedeb.api.services.ngrams_service.mappers.ngrams_to_ngram_result')
     def test_get_ngrams_success(self, mock_to_result, mock_to_opts, mock_n_grams):
         """Test get_ngrams returns NGramResult."""
-        from api_swedeb.api.utils.common_params import CommonQueryParams
-
         mock_corpus = MagicMock()
         mock_to_opts.return_value = [{"cqp": "test"}]
 
@@ -56,8 +54,6 @@ class TestNGramsServiceMethods:
     @patch('api_swedeb.api.services.ngrams_service.mappers.query_params_to_CQP_opts')
     def test_get_ngrams_empty_opts(self, mock_to_opts, mock_n_grams):
         """Test get_ngrams returns empty result when no options."""
-        from api_swedeb.api.utils.common_params import CommonQueryParams
-
         mock_corpus = MagicMock()
         mock_to_opts.return_value = []
 
@@ -76,8 +72,6 @@ class TestNGramsServiceMethods:
     @patch('api_swedeb.api.services.ngrams_service.mappers.query_params_to_CQP_opts')
     def test_get_ngrams_empty_search_term_raises(self, mock_to_opts):
         """Test get_ngrams raises ValueError for empty search terms."""
-        from api_swedeb.api.utils.common_params import CommonQueryParams
-
         mock_corpus = MagicMock()
 
         service = NGramsService()
@@ -98,8 +92,6 @@ class TestNGramsServiceMethods:
     @patch('api_swedeb.api.services.ngrams_service.mappers.ngrams_to_ngram_result')
     def test_get_ngrams_with_list_terms(self, mock_to_result, mock_to_opts, mock_n_grams):
         """Test get_ngrams with list of search terms."""
-        from api_swedeb.api.utils.common_params import CommonQueryParams
-
         mock_corpus = MagicMock()
         mock_to_opts.return_value = [{"cqp": "test"}]
